move source to /src
